Conflicts:
	debugfs/set_fields.c
	tests/f_mmp/script
	tests/f_mmp_garbage/script
	tests/m_mmp/script
	tests/t_mmp_1on/script
	tests/t_mmp_2off/script

+++ lib/ext2fs/tst_libext2fs.c
@@ -37,8 +37,8 @@ static int print_blocks_proc(ext2_filsys fs EXT2FS_ATTR((unused)),
 			     blk64_t ref_block, int ref_offset,
 			     void *private EXT2FS_ATTR((unused)))
 {
-	printf("%6d %8llu (%d %llu)\n", blockcnt, *blocknr,
-	       ref_offset, ref_block);
+	printf("%6lld %8llu (%d %llu)\n", (long long) blockcnt,
+	       (unsigned long long)*blocknr, ref_offset, ref_block);
 	return 0;
 }
